Added useWeb3 hook to Drain Safe app
